backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2025.1.31",
    "dateparser>=1.2.1",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.8",
    "googlesearch-python>=1.3.0",
    "jigsawstack>=0.1.29",
    "keybert>=0.9.0",
    "lxml-html-clean>=0.4.1",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "praw>=7.8.1",
    "snscrape>=0.7.0.20230622",
    "transformers>=4.49.0",
    "uvicorn>=0.34.0",
]
